pyproject.toml
[project]
name = "ARES-agent"
version = "0.1.0"
description = "Introduction to AI agent"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "llama-cpp-python>=0.3.16",
    "mcp[cli]>=1.23.0",
    "pydantic>=2.12.5",
    "requests>=2.32.5",
    "streamlit>=1.51.0",
]
